work on daalink, fop actor, mDo machine, msg_scrn, + various (#201)
* work on fop actor / actor mng, daalink, d_a_obj_item

* d_a_title mostly decompiled

* daalink / d_event / JMessage / dmsg_out_font work

* msg_scrn_base / msg_scrn_boss

* some work on mDo machine, d_menu_save, d_tresure, and various

* remove asm

* progress

+++ libs/JSystem/JUtility/JUTResFont.cpp
@@ -380,7 +380,7 @@ u16 JUTResFont::getCellWidth() const {
 #pragma push
 #pragma optimization_level 0
 #pragma optimizewithasm off
-asm u16 JUTResFont::getCellWidth() const {
+asm int JUTResFont::getCellWidth() const {
     nofralloc
 #include "asm/JSystem/JUtility/JUTResFont/getCellWidth__10JUTResFontCFv.s"
 }
